streamlit_app/requirements.txt.txt
streamlit
pandas
snowflake-connector-python
plotly
numpy
pyyaml
confluent-kafka
streamlit-autorefresh
